requirements.txt
pandas
matplotlib
seaborn
numpy 
scikit-learn
opencv-python
tensorflow
plotly
streamlit
google-generativeai
keras
